Add files via upload
Weather data can be as unpredictable as a cat on a hot tin roof. But with the right analysis, we can uncover patterns and make predictions that are as reliable as a Swiss watch. In this notebook, I have dived into a weather dataset to explore the intricacies of climate data.