Add files via upload
This project demonstrates the use of Simple Linear Regression to predict brain weight based on head size. The relationship between the two variables is analyzed using a dataset that contains measurements of head size (in cm³) and brain weight (in grams) of individuals.

The model is trained on the data to learn a linear relationship, allowing it to make predictions of brain weight when given a head size. The primary goal is to explore the correlation between head size and brain weight and use regression analysis to quantify this relationship.